Implemented working notifications
Notifications now working - They appear on the home screen when the user has saved their address, enabled notifications and their next collection is either today or tomorrow.

Added an error to Collection Dates - If the user tries to enter the collection dates screen without having set an address, an error is now displayed guiding them to set an address first in the settings menu.